pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiocron>=2.1",
    "chromadb>=1.1.1",
    "duckdb>=1.2.2",
    "faster-whisper>=1.1.1",
    "fortune-python>=1.1.1",
    "garminconnect>=0.2.26",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.14",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.18",
    "langchain-text-splitters>=0.3.11",
    "loguru>=0.7.3",
    "ollama>=0.4.8",
    "openai-agents>=0.0.14",
    "pandas>=2.2.3",
    "polars>=1.29.0",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "scipy>=1.16.2",
    "sentence-transformers>=5.1.1",
]

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 120
combine_as_imports = true

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "mypy>=1.16.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
    "types-pyyaml>=6.0.12.20250516",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
